# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  displayName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface PublicUser {
  id: number;
  email: string;
  displayName: string;
  createdAt: string;
}

export interface GithubLinkInfo {
  githubUsername: string;
  githubEmail?: string | null;
  linkedAt: string;
}

export interface AuthResponse {
  user: PublicUser;
}

export interface MeResponse {
  user: PublicUser;
  github: GithubLinkInfo | null;
}

export interface Storyworld {
  id: number;
  repoOwner: string;
  repoName: string;
  title: string;
  stewardId?: number | null;
  canonBranchRef: string;
  createdAt: string;
  updatedAt: string;
}

export type StoryPathState =
  (typeof StoryPathState)[keyof typeof StoryPathState];

export const StoryPathState = {
  personal: "personal",
  open: "open",
  proposed: "proposed",
  "published-alternate": "published-alternate",
} as const;

export interface StoryPath {
  id: number;
  storyworldId: number;
  branchRef: string;
  title: string;
  originPathId?: number | null;
  state: StoryPathState;
  createdAt: string;
  updatedAt: string;
}

export interface Contribution {
  id: number;
  storyworldId: number;
  pathId: number;
  commitSha: string;
  contributorId?: number | null;
  title: string;
  summary?: string | null;
  createdAt: string;
}

export interface TranscribeRequest {
  /** Base64-encoded audio data */
  audioBase64: string;
  /** MIME type of the audio (e.g. audio/m4a, audio/webm) */
  mimeType: string;
}

export interface TranscribeResponse {
  /** Transcribed text from the audio */
  text: string;
}

export type ProposalState = (typeof ProposalState)[keyof typeof ProposalState];

export const ProposalState = {
  draft: "draft",
  submitted: "submitted",
  "under-review": "under-review",
  "returned-with-notes": "returned-with-notes",
  "accepted-into-canon": "accepted-into-canon",
  "published-alternate": "published-alternate",
} as const;

export interface Proposal {
  id: number;
  storyworldId: number;
  pathId: number;
  prNumber: number;
  state: ProposalState;
  submittedAt: string;
  decidedAt?: string | null;
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Missing or invalid fields
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Not authenticated
 */
export type UnauthorizedResponse = ErrorResponse;

export type Logout200 = {
  ok: boolean;
};

export type GithubCallbackParams = {
  code?: string;
  state?: string;
  error?: string;
};