# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storyworlds/{id}:
    get:
      operationId: getStoryworld
      tags: [storyworlds]
      summary: Get a storyworld by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Storyworld
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Storyworld"
        "404":
          $ref: "#/components/responses/NotFound"
  /storyworlds/{id}/paths:
    get:
      operationId: listStoryPaths
      tags: [storyworlds]
      summary: List story paths for a storyworld
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Story path list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoryPath"
        "404":
          $ref: "#/components/responses/NotFound"
  /storyworlds/{id}/paths/{pathId}/contributions:
    get:
      operationId: listContributions
      tags: [storyworlds]
      summary: List contributions for a story path
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pathId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contribution list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contribution"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Storyworld:
      type: object
      properties:
        id:
          type: integer
        repoOwner:
          type: string
        repoName:
          type: string
        title:
          type: string
        stewardId:
          type: integer
          nullable: true
        canonBranchRef:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - repoOwner
        - repoName
        - title
        - canonBranchRef
        - createdAt
        - updatedAt
    StoryPath:
      type: object
      properties:
        id:
          type: integer
        storyworldId:
          type: integer
        branchRef:
          type: string
        title:
          type: string
        originPathId:
          type: integer
          nullable: true
        state:
          type: string
          enum: [personal, open, proposed, published-alternate]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - storyworldId
        - branchRef
        - title
        - state
        - createdAt
        - updatedAt
    Contribution:
      type: object
      properties:
        id:
          type: integer
        storyworldId:
          type: integer
        pathId:
          type: integer
        commitSha:
          type: string
        contributorId:
          type: integer
          nullable: true
        title:
          type: string
        summary:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storyworldId
        - pathId
        - commitSha
        - title
        - createdAt
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...